pyproject.toml
[project]
name = "argo-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "opentelemetry-exporter-otlp-proto-grpc>=1.41.0",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
    "opentelemetry-sdk>=1.41.0",
]
